refactor: audit remediation batch F (PR1) - web layer structure (2.10.29)
Splits web/config_app.py into one module per config screen (connections,
users, libraries, settings) and utils/plex.py into an API adapter plus a
rating/label policy module, replaces 41 repeated field-error blocks with
one Jinja macro, and brings every touched module's docstring coverage up
to parity with the rest of the codebase.
